It turns out that we DO have a test for op.CALL opimpl: test_call_instance.
The test passed "by chance" in doppler mode because it didn't actually
do any redshift, and the interpreter was able to execute the blue call
anyway. But if we add an assert to ensure that the result of the
redshift is a const, the the test would fail, and the "if call in
self.opimpl" condition makes it passing.