Big simplification in how W_Func are typechecked and called.
Until now, we had two complete separate paths for calling functions
and calling opimpls, now we unify them.

Introduce W_Func.op_CALL which returns the OpImpl for calling a
function. It is still a special case but only inside the ASTFrame (see
W_DirectCall.__doc__), and it lets us to kill a lot of unnecessary
code.

This is still WIP as we need to improve error messages and fix the
corresponding tests.