Fix "ptr[T]" in the SPy backend.
This was broken needed because types are "special" blue expressions
which are directly evaluated by astframe instead of by doppler, so
they were bypassing the logic in doppler.make_const.

The solution is to move the logic to assign FQNs from
doppler.make_const to vm.make_fqn_const, and call it from astframe.