Fix a nasty bug in the ScopeAnalyzer.
The scope analyzer always assumed that if a symbol is present in a
symtable, then it's defined there. This assumption is valid during the
"declare()" phase, but no longer valid after "flatten()" phase.

The consequence is that in case of deeply nested closures, sometimes
we tried to find a symbol in an outer scope where the name was used,
instead of the scope where the name is DEFINED.