Make blue W_OpArgs when we call things like vm.getitem, vm.eq, etc.
This is needed mainly because we call vm.getitem to create a new
specialized type, and thus now we require args to be blue (equivalent
of doing `list[T]`... T must be blue).

I'm not 100% sure that making them blue if ALWAYS the right choice,
but I suppose that for now it's good enough.