Refactor how SPy tracebacks are collected.
The old logic to keep track for `.loc` in ASTFrame was buggy. Instead,
we reconstruct the precise loc of each SPyFrame by inspecting the
frames associated to eval_expr and exec_stmt.

Moreover, kill for now the possibility of extracting the stack without
raising an exception. It's possible, but not worth the complexity now.