Turn ArgSpec.Convert.w_conv into w_conf_opimpl.
w_conv was a function and we always called it only with a single
argument (the argument which needs the conversion).

By swithing to OpImpl, we can encode more complex behavior and
propagate W_MetaArgs from above, making it possible for
__CONVERT_FROM__ and __CONVERT_TO__ to return complex OpSpec with a
custom list of metaargs.

test_basic::test_function_call_conversion works.